Move external libraries to util folder
